[Feature #108847854] rename messages app to conversations because of naming conflict with django.contrib.messages
