+++ troupon/conversations/tests/test_models.py
@@ -1,4 +1,4 @@
-from ..models import Message
+from conversations.models import Message
 from django.test import TestCase
 from django.contrib.auth.models import User
